requirements-train.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
albumentations>=1.3.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
onnx>=1.14.0
onnxruntime>=1.15.0
Pillow>=9.5.0
tqdm>=4.65.0
tensorboard>=2.13.0
kaggle>=1.5.0
imagehash>=4.3.0
